Refactored code such that records.updated metric applies only to BQ Sink and not BQ Multi Sink
